backend/requirements.txt
asgiref==3.9.1
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
cryptography==45.0.6
Django==5.2.5
djangorestframework==3.16.1
ecdsa==0.19.1
idna==3.10
pyasn1==0.6.1
pycparser==2.22
python-dotenv==1.1.1
python-jose==3.5.0
requests==2.32.4
rsa==4.9.1
six==1.17.0
sqlparse==0.5.3
urllib3==2.5.0
